libs/deepagents-telegram/pyproject.toml
[project]
name = "deepagents-telegram"
version = "0.1.0"
description = "Telegram bot interface for DeepAgents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
    "deepagents",
    "deepagents-cli",
]

[project.scripts]
deepagents-telegram = "deepagents_telegram.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["deepagents_telegram"]
